Move BenchmarkResult to new file
