+++ src/hyperfine/export/asciidoc.rs
@@ -1,7 +1,7 @@
 use super::Exporter;
 
+use crate::hyperfine::benchmark_result::BenchmarkResult;
 use crate::hyperfine::format::format_duration_value;
-use crate::hyperfine::types::BenchmarkResult;
 use crate::hyperfine::units::Unit;
 
 use std::io::Result;
