+++ src/hyperfine/export/csv.rs
@@ -1,6 +1,6 @@
 use super::Exporter;
 
-use crate::hyperfine::types::BenchmarkResult;
+use crate::hyperfine::benchmark_result::BenchmarkResult;
 use crate::hyperfine::units::Unit;
 
 use std::borrow::Cow;
